Move `DECONSTRUCT` into `wire-common`.
