+++ packages/data-model/src/fabric-instances/BaseFabricInstance.ts
@@ -1,4 +1,8 @@
-import { FabricInstance } from "../interface.ts";
+import { FabricInstance, type FabricValue } from "../interface.ts";
+import {
+  DECONSTRUCT,
+  FabricDeconstructable,
+} from "../wire-common/interface.ts";
 
 /**
  * Abstract base class providing shared scaffolding for `FabricInstance`
@@ -12,10 +16,14 @@ import { FabricInstance } from "../interface.ts";
  * `shallowClone()`/`shallowUnfrozenClone()` template-method split), at which
  * point individual subclasses stop implementing `deepClone()` directly.
  */
-export abstract class BaseFabricInstance extends FabricInstance {
+export abstract class BaseFabricInstance extends FabricInstance
+  implements FabricDeconstructable {
   //
   // Instance members
   //
+
+  /** @inheritDoc */
+  abstract [DECONSTRUCT](): FabricValue;
 
   /**
    * The type tag to use for this instance to identify it in wire protocols.
@@ -54,7 +62,7 @@ export abstract class BaseFabricInstance extends FabricInstance {
    * the fully abstract `FabricInstance`. Throws a "shouldn't happen" error if
    * there's trouble.
    */
-  static wireTypeTagOf(value: FabricInstance): string {
+  static wireTypeTagOf(value: FabricInstance | FabricDeconstructable): string {
     if (!(value instanceof BaseFabricInstance)) {
       throw new Error(
         "Shouldn't happen: Encountered a `FabricInstance` which is not a `BaseFabricInstance`.",
