+++ packages/data-model/src/json-wire/FabricInstanceHandler.ts
@@ -1,5 +1,9 @@
-import { DECONSTRUCT, FabricInstance, type FabricValue } from "../interface.ts";
-import type { ReconstructionContext } from "../wire-common/interface.ts";
+import { FabricInstance, type FabricValue } from "../interface.ts";
+import {
+  DECONSTRUCT,
+  type FabricDeconstructable,
+  type ReconstructionContext,
+} from "../wire-common/interface.ts";
 import { ExplicitTagValue } from "../fabric-instances/ExplicitTagValue.ts";
 import type {
   JsonWireValue,
@@ -42,7 +46,7 @@ export const FabricInstanceHandler: TypeHandler = {
     codec: TypeHandlerCodec,
     recurse: (v: FabricValue) => JsonWireValue,
   ): JsonWireValue {
-    const inst = value as FabricInstance;
+    const inst = value as FabricDeconstructable;
 
     // For `ExplicitTagValue`, use the preserved original `wireTypeTag` and
     // `state`.
