+++ packages/data-model/test/fabric-instances/FabricSet.test.ts
@@ -1,12 +1,12 @@
 import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
 import {
-  DECONSTRUCT,
   DEEP_FREEZE,
   FabricInstance,
   type FabricValue,
   IS_DEEP_FROZEN,
 } from "../../src/interface.ts";
+import { DECONSTRUCT } from "../../src/wire-common/interface.ts";
 import { FabricSet } from "../../src/fabric-instances/FabricSet.ts";
 import { FrozenSet } from "../../src/frozen-builtins.ts";
 import { deepFreeze, isDeepFrozenFabricValue } from "../../src/deep-freeze.ts";
