+++ packages/data-model/test/native-conversion.test.ts
@@ -1,13 +1,12 @@
 import { describe, it } from "@std/testing/bdd";
 import { expect } from "@std/expect";
 import {
-  DECONSTRUCT,
   DEEP_FREEZE,
   FabricInstance,
   type FabricValue,
   IS_DEEP_FROZEN,
 } from "../src/interface.ts";
-import { RECONSTRUCT } from "../src/wire-common/interface.ts";
+import { DECONSTRUCT, RECONSTRUCT } from "../src/wire-common/interface.ts";
 import { FabricError } from "../src/fabric-instances/FabricError.ts";
 import { FabricMap } from "../src/fabric-instances/FabricMap.ts";
 import { FabricSet } from "../src/fabric-instances/FabricSet.ts";
